Add explicit casts for narrowing conversions and force LF in format.ts
Cast int/float/DWORD/size_t conversions flagged by
bugprone-narrowing-conversions under src/. After clang-format, cmd/format.ts
rewrites files to LF so Windows runs do not leave CRLF.

prompt: fix bugprone-narrowing-conversions under src/; fix format.ts to ensure LF

fix build

+++ src/Annotation.cpp
@@ -832,7 +832,7 @@ static void PdfColorToFloat(PdfColor c, float rgb[3]) {
 
 static float GetOpacityFloat(PdfColor c) {
     u8 alpha = GetAlpha(c);
-    return alpha / 255.0f;
+    return (float)alpha / 255.0f;
 }
 
 static PdfColor MkPdfColorFromFloat(float rf, float gf, float bf) {
